Rename the results module to protocol
