+++ engine/vault/src/vault/protocol.rs
